Move MetricInputScreen to dedicated module
