+++ ui/screens/__init__.py
@@ -1,0 +1,3 @@
+from .metric_input_screen import MetricInputScreen
+
+__all__ = ["MetricInputScreen"]
